Fix pmd warnings
