+++ backend/src/test/java/at/uibk/dps/rm/handler/function/RuntimeTemplateHandlerTest.java
@@ -2,7 +2,6 @@ package at.uibk.dps.rm.handler.function;
 
 import at.uibk.dps.rm.exception.NotFoundException;
 import at.uibk.dps.rm.service.rxjava3.database.function.RuntimeService;
-import at.uibk.dps.rm.testutil.mockprovider.Mockprovider;
 import at.uibk.dps.rm.testutil.objectprovider.TestFunctionProvider;
 import at.uibk.dps.rm.util.serialization.JsonMapperConfig;
 import io.reactivex.rxjava3.core.Single;
@@ -23,6 +22,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
+import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.when;
 
 /**
@@ -70,7 +70,7 @@ public class RuntimeTemplateHandlerTest {
         when(rc.pathParam("id")).thenReturn(String.valueOf(runtimeId));
         when(runtimeService.findOne(runtimeId)).thenReturn(Single.just(runtime));
 
-        try (MockedStatic<Vertx> mockedVertx = Mockprovider.mockVertx()) {
+        try (MockedStatic<Vertx> mockedVertx = mockStatic(Vertx.class)) {
             when(context.owner()).thenReturn(vertx);
             when(vertx.fileSystem()).thenReturn(fileSystem);
             when(fileSystem.exists(templatePath)).thenReturn(Single.just(true));
@@ -112,7 +112,7 @@ public class RuntimeTemplateHandlerTest {
 
         when(rc.pathParam("id")).thenReturn(String.valueOf(runtimeId));
         when(runtimeService.findOne(runtimeId)).thenReturn(Single.just(runtime));
-        try (MockedStatic<Vertx> mockedVertx = Mockprovider.mockVertx()) {
+        try (MockedStatic<Vertx> mockedVertx = mockStatic(Vertx.class)) {
             when(context.owner()).thenReturn(vertx);
             when(vertx.fileSystem()).thenReturn(fileSystem);
             when(fileSystem.exists(templatePath)).thenReturn(Single.just(false));
@@ -137,7 +137,7 @@ public class RuntimeTemplateHandlerTest {
         when(rc.pathParam("id")).thenReturn(String.valueOf(runtimeId));
         when(runtimeService.findOne(runtimeId)).thenReturn(Single.just(runtime));
 
-        try (MockedStatic<Vertx> mockedVertx = Mockprovider.mockVertx()) {
+        try (MockedStatic<Vertx> mockedVertx = mockStatic(Vertx.class)) {
             when(context.owner()).thenReturn(vertx);
             when(vertx.fileSystem()).thenReturn(fileSystem);
             when(fileSystem.exists(templatePath)).thenReturn(Single.just(true));
